evals/pyproject.toml
[project]
name = "evals"
version = "0.2.0"
description = "DeepEval tests for the Upsun CLI skill"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepeval>=3.8.2",
    "google-genai>=1.60.0",
    "pytest-html>=4.1.1",
]
